Fix bug where opponent doesn't find out unit is dead
Seems to be related to a unit that's still moving but is killed before
the moevement fully complete's (or something like that). The
"fix", without fully understanding the problem, is to mark both both
tiles (previous and current) for the killed unit as dirty, so the
killing player will be sure to get the message that the unit is dead.

Otherwise the killing player thinks there's an invincible non-moving
unit hanging out on a tile forever.